docs: add detail component
